requirements.txt
numpy
scipy
pandas
matplotlib
altair
metaflow
python-dotenv
tqdm
filelock
typer
sh
cdlib ==0.2.5
tensorflow ==2.7.0
torch==1.9.0
sentencepiece==0.1.96
faiss-cpu==1.7.1.post2
googletrans==4.0.0-rc1
networkx ==2.6.3
sentence-transformers ==2.0.0
xlsxwriter==3.0.3
langdetect ==1.0.9
openpyxl==3.0.9
sentence_transformers
sklearn
stop_words
langdetect
textwrap3
umap-learn
cairocffi
leidenalg
igraph
